Improves message formats

+++ bestia-zoneserver/src/main/kotlin/net/bestia/zoneserver/entity/component/TemperatureComponent.kt
@@ -9,5 +9,5 @@ data class TemperatureComponent(
     override val entityId: Long,
     val minTolerableTemperature: Int,
     val maxTolerableTemperature: Int,
-    val currentTemperature: Int
+    val currentTemperature: Float
 ) : Component
